requirements.txt
numpy   2.2.5
pillow  11.2.1
pip     25.1
psutil  7.0.0
pygame  2.6.1
pynput  1.8.1
six     1.17.0